fix sonar rule "Standards outputs..."

+++ src/main/java/tn/esprit/spring/services/EmployeServiceImpl.java
@@ -1,9 +1,13 @@
 package tn.esprit.spring.services;
 
+import java.text.MessageFormat;
 import java.util.ArrayList;
+
 import java.util.Date;
 import java.util.List;
+import java.util.Optional;
 
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -19,7 +23,9 @@ import tn.esprit.spring.repository.DepartementRepository;
 import tn.esprit.spring.repository.EmployeRepository;
 import tn.esprit.spring.repository.TimesheetRepository;
 
+
 @Service
+@Slf4j
 public class EmployeServiceImpl implements IEmployeService {
 
 	@Autowired
@@ -43,17 +49,35 @@ public class EmployeServiceImpl implements IEmployeService {
 	}
 
 
-	public void mettreAjourEmailByEmployeId(String email, int employeId) {
-		Employe employe = employeRepository.findById(employeId).get();
-		employe.setEmail(email);
-		employeRepository.save(employe);
+	 public void mettreAjourEmailByEmployeId(String email, int employeId) {
+	        Optional<Employe> empl = employeRepository.findById(employeId);
+	        Employe employe = null;
+	        if (empl.isPresent()) {
+	            employe = empl.get();
+	            log.info(MessageFormat.format("Employe Id,{0}", employe.getId()));
+	            employe.setEmail(email);
+	            log.info(MessageFormat.format("Employe Id   {0}", employe));
+	            employeRepository.save(employe);
+	        } else {
+	            log.info(MessageFormat.format("Employe doesn''t exist{0}", employe));
+	            }
+	        }
 
-	}
 
-	@Transactional	
-	public void affecterEmployeADepartement(int employeId, int depId) {
-		Departement depManagedEntity = deptRepoistory.findById(depId).get();
-		Employe employeManagedEntity = employeRepository.findById(employeId).get();
+	 @Transactional
+	    public void affecterEmployeADepartement(int employeId, int depId) {
+	        log.info(MessageFormat.format("Department Id {0}", depId));
+	        log.info(MessageFormat.format("Employe Id {0}", employeId));
+	        Departement depManagedEntity = null;
+	        Employe employeManagedEntity = null;
+	        Optional<Employe> empl = employeRepository.findById(employeId);
+	        Optional<Departement> dep = deptRepoistory.findById(depId);
+	        if (dep.isPresent() && empl.isPresent()) {
+	            depManagedEntity = dep.get();
+	            log.debug(MessageFormat.format("Department Id  {0}", depId));
+	            employeManagedEntity = empl.get();
+	            log.debug(MessageFormat.format("Employe Id, {0}", employeId));
+
 
 		if(depManagedEntity.getEmployes() == null){
 
@@ -65,22 +89,29 @@ public class EmployeServiceImpl implements IEmployeService {
 			depManagedEntity.getEmployes().add(employeManagedEntity);
 		}
 
-		// à ajouter? 
+	
 		deptRepoistory.save(depManagedEntity); 
+		}
 
 	}
-	@Transactional
-	public void desaffecterEmployeDuDepartement(int employeId, int depId)
-	{
-		Departement dep = deptRepoistory.findById(depId).get();
-
+	 
+	 @Transactional
+	    public void desaffecterEmployeDuDepartement(int employeId, int depId) {
+	        log.info(MessageFormat.format("Department Id{0}", depId));
+	        log.info(MessageFormat.format("Employe Id , {0}", employeId));
+	        Departement dep = null;
+	        Optional<Employe> empl = employeRepository.findById(employeId);
+	        Optional<Departement> departement = deptRepoistory.findById(employeId);
+	        if (departement.isPresent() && empl.isPresent()) {
+	            dep = departement.get();
+	            log.debug(MessageFormat.format("Department Id{0}", depId));
 		int employeNb = dep.getEmployes().size();
 		for(int index = 0; index < employeNb; index++){
 			if(dep.getEmployes().get(index).getId() == employeId){
 				dep.getEmployes().remove(index);
-				break;//a revoir
+				break;
 			}
-		}
+		}}
 	} 
 	
 	// Tablesapce (espace disque) 
@@ -90,39 +121,67 @@ public class EmployeServiceImpl implements IEmployeService {
 		return contrat.getReference();
 	}
 
-	public void affecterContratAEmploye(int contratId, int employeId) {
-		Contrat contratManagedEntity = contratRepoistory.findById(contratId).get();
-		Employe employeManagedEntity = employeRepository.findById(employeId).get();
+	 public void affecterContratAEmploye(int contratId, int employeId) {
+	        log.info(MessageFormat.format("Employe Id {0}", employeId));
+	        log.info(MessageFormat.format("Contrat Id {0}", contratId));
+	        Contrat contratManagedEntity = null;
+	        Employe employeManagedEntity = null;
+	        Optional<Employe> empl = employeRepository.findById(employeId);
+	        Optional<Contrat> contrat = contratRepoistory.findById(employeId);
+	        if (contrat.isPresent() && empl.isPresent()) {
+	            contratManagedEntity = contrat.get();
+	            log.debug(MessageFormat.format("Contrat Id{0}", contratId));
+	            employeManagedEntity = empl.get();
+	            log.debug(MessageFormat.format("Employe Id{0}", employeId));
 
 		contratManagedEntity.setEmploye(employeManagedEntity);
 		contratRepoistory.save(contratManagedEntity);
-
-	}
-
-	public String getEmployePrenomById(int employeId) {
-		Employe employeManagedEntity = employeRepository.findById(employeId).get();
-		return employeManagedEntity.getPrenom();
-	}
-	 
-	public void deleteEmployeById(int employeId)
-	{
-		Employe employe = employeRepository.findById(employeId).get();
-
-		//Desaffecter l'employe de tous les departements
-		//c'est le bout master qui permet de mettre a jour
-		//la table d'association
-		for(Departement dep : employe.getDepartements()){
-			dep.getEmployes().remove(employe);
 		}
 
-		employeRepository.delete(employe);
 	}
 
-	public void deleteContratById(int contratId) {
-		Contrat contratManagedEntity = contratRepoistory.findById(contratId).get();
-		contratRepoistory.delete(contratManagedEntity);
+	 public String getEmployePrenomById(int employeId) {
+	        Employe employe = null;
+	        Optional<Employe> empl = employeRepository.findById(employeId);
+	        if (empl.isPresent()) {
+	            employe = empl.get();
+	            return employe.getPrenom();
+	        } else {
+	            log.info("Employe doesn't exist");
+	            return null;
+	        }
+	    }
+	 
+	 public void deleteEmployeById(int employeId) {
+	        Employe employe = null;
+	        Optional<Employe> empl = employeRepository.findById(employeId);
+	        if (empl.isPresent()) {
+	            employe = empl.get();
+	            //Desaffecter l'employe de tous les departements
+	            //c'est le bout master qui permet de mettre a jour
+	            //la table d'association
+	            for (Departement dep : employe.getDepartements()) {
+	                dep.getEmployes().remove(employe);
+	            }
+	            employeRepository.delete(employe);
+	        } else {
+	            log.info("Employe doesn't exist");
+	        }
+	    }
 
-	}
+
+	  public void deleteContratById(int contratId) {
+	        log.info(MessageFormat.format("Contrat Id{0}", contratId));
+	        Optional<Contrat> contrat = contratRepoistory.findById(contratId);
+	        Contrat contratManagedEntity = null;
+	        if (contrat.isPresent()) {
+	            contratManagedEntity = contrat.get();
+	            contratRepoistory.delete(contratManagedEntity);
+	            log.info(MessageFormat.format("Contrat Id deleted{0}", contratId));
+	        } else {
+	            log.info("Contrat doesn't exist");
+	        }
+	    }
 
 	public int getNombreEmployeJPQL() {
 		return employeRepository.countemp();
